refactor: 패키지 의존성 1차 수정
