+++ src/main/java/back/ecommerce/common/CustomLogger.java
@@ -1,4 +1,4 @@
-package back.ecommerce.controller.common;
+package back.ecommerce.common;
 
 import org.slf4j.event.Level;
