+++ src/main/java/back/ecommerce/repository/product/ProductQueryDslRepository.java
@@ -1,7 +1,7 @@
 package back.ecommerce.repository.product;
 
 import static back.ecommerce.domain.product.QProduct.*;
-import static back.ecommerce.dto.request.product.ProductSortCondition.*;
+import static back.ecommerce.domain.condition.ProductSortCondition.*;
 import static com.querydsl.core.types.dsl.Expressions.*;
 
 import java.util.ArrayList;
@@ -17,8 +17,8 @@ import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 
 import back.ecommerce.domain.product.Category;
-import back.ecommerce.dto.request.product.ProductSearchCondition;
-import back.ecommerce.dto.request.product.ProductSortCondition;
+import back.ecommerce.domain.condition.ProductSearchCondition;
+import back.ecommerce.domain.condition.ProductSortCondition;
 import back.ecommerce.dto.response.product.ProductDto;
 import lombok.RequiredArgsConstructor;
