+++ src/main/java/back/ecommerce/service/product/ProductService.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Service;
 
 import back.ecommerce.domain.product.Category;
 import back.ecommerce.dto.response.product.ProductDto;
-import back.ecommerce.dto.request.product.ProductSearchCondition;
+import back.ecommerce.domain.condition.ProductSearchCondition;
 import back.ecommerce.dto.response.product.ProductListResponse;
 import back.ecommerce.repository.product.ProductQueryDslRepository;
 import lombok.RequiredArgsConstructor;
