+++ src/test/java/back/ecommerce/repository/ProductRepositoryTest.java
@@ -21,7 +21,7 @@ import back.ecommerce.config.jpa.JpaAuditingConfig;
 import back.ecommerce.constant.PageConstant;
 import back.ecommerce.domain.product.Category;
 import back.ecommerce.domain.product.Product;
-import back.ecommerce.dto.request.product.ProductSearchCondition;
+import back.ecommerce.domain.condition.ProductSearchCondition;
 import back.ecommerce.dto.response.product.ProductDto;
 import back.ecommerce.repository.product.ProductQueryDslRepository;
 import back.ecommerce.repository.product.ProductRepository;
